rename `editable` -> `text`, `blockeditable` -> `blocktext`
