+++ text/__init__.py
@@ -1,2 +1,2 @@
-VERSION = (1, 5, 6)
+VERSION = (1, 6, 0)
 __version__ = '.'.join((str(i) for i in VERSION))
